Implemented prefix adder from page 242 of RISCV Harris-Harris
Note: The Gi:j and Pi:j equations make solving this problem abnormally
difficult. The equations and the k value do not map well to how the
hardware is described in the book.

I might not being something simple or intuitive, but, the k and k-1
notation creates variable solutions which are not what is shown in the
book itself.